setup todos and home rooting
